Add a helper function to read/update json from local storage
Read a key from localStorage parse it as a json and pass it
to a custom function, if the function returns a value, this value is
stringified and used as a new value for the key in localStorage.
it is responsability of the custom function to check for nulls and if
it is the case initialize the key.

Issue edina/fieldtrip-gb#244